requirements.txt
Django==1.9.5
Markdown==2.6.6
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.23
celery==3.1.23
django-filter==0.13.0
djangorestframework==3.3.3
httplib2==0.9.2
kombu==3.0.35
pytz==2016.4
six==1.10.0
twilio==5.4.0
wsgiref==0.1.2
